+++ modules/infra/outputs.tf
@@ -1,3 +1,13 @@
+output "aad_pod_identity_namespace" {
+  description = "Namespace used for AAD Pod Identity"
+  value       = kubernetes_namespace.add_pod_identity.metadata.0.name
+}
+
+output "aad_pod_identity_azure_identity" {
+  description = "Identity object for AAD Pod Identity"
+  value       = azurerm_user_assigned_identity.aad_pod_identity
+}
+
 output "aad_pod_identity_id" {
   value = azurerm_user_assigned_identity.aad_pod_identity.id
 }