+++ modules/infra/variables.tf
@@ -22,6 +22,12 @@ variable "location" {
   type        = string
 }
 
+variable "aadpodidentity_chart_repository" {
+  description = "URL of the Helm chart repository"
+  type        = string
+  default     = "https://raw.githubusercontent.com/Azure/aad-pod-identity/master/charts"
+}
+
 variable "aadpodidentity_chart_version" {
   description = "Azure Active Directory Pod Identity Chart version"
   type        = string