+++ modules/tools/cert-manager/versions.tf
@@ -1,0 +1,13 @@
+terraform {
+  required_version = "> 0.12.26"
+  required_providers {
+    helm = {
+      source  = "hashicorp/helm"
+      version = "1.1.1"
+    }
+    kubernetes = {
+      source  = "hashicorp/kubernetes"
+      version = ">= 1.11.1"
+    }
+  }
+}