+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# Unreleased
+
+Added
+ * AZ-377: Core tools as documented sub-module
+
+Updated
+ * AZ-381: Harmonize variables and add outputs
+ * AZ-380: AKS module: pin velero azure image
+ * AZ-379: Fix kured chart repository
+ * AZ-382: Update of Kured chart version
+ 
 # v3.2.0 - 2020-10-20
 
 Fixed